For NCTU OSDI course, lab1 together with lab2
lab1 is for kernel debugging, we try to use some tools such as gdb, cscope, ctags... to find dugs in the kernel code

lab2 is for multiple booting, from this lab, we learned the flow of OS kernel booting.
Besieds, with int 0x16 software interrupt, the system will get the keystroke ASCii number into the AL register.